Multiple bug fixes, most for PID flow control, but one for the menus as well where the last digit of an entry with divisor 1 could not be modified, removed PWM 8K code (the interrupt just wouldnt be fast enough to handle low and high % duty cycles, we'd blow an interrupt and invert out logic). Modified PID flow control to use a 10hz output controlled by the PID by timer code. Added a manual mode for PID flow control in case PID just wont work due to the long sample time required.
git-svn-id: http://brewtroller.googlecode.com/svn/trunk/BrewTroller@690 786cebc4-e6f9-11dd-9c15-fdd50ba39c1e

+++ EEPROM.pde
@@ -293,7 +293,7 @@ void setVolCalib(byte vessel, byte slot, unsigned int value, unsigned long vol) 
 //**********************************************************************************
 //setpoints (299-301)
 //**********************************************************************************
-void setSetpoint(byte vessel, byte value) {
+void setSetpoint(byte vessel, int value) {
   #if defined PID_FLOW_CONTROL || defined USESTEAM
     if (vessel == VS_STEAM) setpoint[vessel] = value;
     else setpoint[vessel] = value * SETPOINT_MULT;
